+++ content/chronicles/orlando.mdx
@@ -1,0 +1,32 @@
+---
+title: "orlando"
+date: "2026-07-04"
+summary: "Take 2; Day 165, Where's Shaq & Penny?"
+draft: false
+infinityStone: false
+---
+
+<ReleaseSection alterEgo="unclejimmy" divider={false}>
+    Hopped on a couple of flights with squad today. 
+    <FolderImageCarousel folder="orlando/squad" />
+
+</ReleaseSection>
+
+<ReleaseSection alterEgo="mark2" releaseId="174">
+    For the past few months my environment has been incredibly flakey again. I'm building out this thing, it sprawls, I understand it, but not enough to know where I might be burning myself short or long term. It hasn't been much more than a minor nuisance and yet with a bit of extra time on this flight, we took a crack at fixing it.
+
+    It sure seems like the homie <PersonTag tag="cipher" /> fixed the issue in a few minutes. I shoulda done so sooner it turns out......also, I haven't tested it yet and I worry about bricking my environment because I've done so in the past.
+
+    ......testing......
+
+    It appears to be a success, along with an unexpected behavior change, robot tells me it knows how to fix that too......testing......
+
+    Dudes, plane landed.
+
+    Well, not this plane, it's still in the air. You get the point.
+
+    ---
+
+
+
+</ReleaseSection>
