+++ package.json
@@ -8,9 +8,10 @@
   "type": "module",
   "scripts": {
     "dev": "next dev",
-    "dev:full": "concurrently -k -n next,content \"npm run dev\" \"npm run content:dev\"",
+    "dev:full": "cross-env NEXT_WATCH_CONTENTLAYER=1 concurrently -k -n next,content \"npm run dev:stable\" \"npm run content:dev\"",
+    "dev:full:turbo": "concurrently -k -n next,content \"npm run dev\" \"npm run content:dev\"",
     "dev:poll": "cross-env WATCHPACK_POLLING=true WATCHPACK_POLLING_INTERVAL=1000 next dev",
-    "dev:stable": "cross-env WATCHPACK_POLLING=true WATCHPACK_POLLING_INTERVAL=1000 NEXT_DISABLE_FS_CACHE=1 next dev",
+    "dev:stable": "cross-env WATCHPACK_POLLING=true WATCHPACK_POLLING_INTERVAL=1000 NEXT_DISABLE_FS_CACHE=1 next dev --webpack",
     "dev:with-contentlayer": "cross-env ENABLE_CONTENTLAYER=1 next dev",
     "dev:turbo": "next dev --turbo",
     "predev": "npm run gen:build-info",
@@ -86,9 +87,11 @@
     "sanity:db": "node scripts/db_sanity.mjs",
     "check:use-server": "bash scripts/check-use-server.sh",
     "gen:build-info": "node scripts/gen-build-info.mjs",
+    "patch:next-breadcrumb-types": "node scripts/patch-next-breadcrumb-types.mjs",
     "gen:images-manifest": "node scripts/gen-optimus-images-manifest.mjs",
     "content:build": "contentlayer2 build",
-    "content:dev": "contentlayer2 dev --clearCache",
+    "content:dev": "contentlayer2 dev",
+    "content:dev:reset": "contentlayer2 dev --clearCache",
     "contentlayer:build": "contentlayer2 build",
     "prepare:content": "npm run gen:build-info && npm run gen:images-manifest && npm run content:build",
     "prepare": "husky",
@@ -157,7 +160,6 @@
     "@fontsource-variable/jetbrains-mono": "^5.2.6",
     "@mdx-js/react": "^3.1.0",
     "@next/bundle-analyzer": "16.1.3",
-    "@next/eslint-plugin-next": "^15.5.2",
     "@next/mdx": "16.1.3",
     "@playwright/test": "^1.56.1",
     "@secretlint/secretlint-rule-preset-recommend": "^11.2.4",
@@ -180,7 +182,7 @@
     "dotenv": "^17.2.1",
     "dotenv-cli": "^10.0.0",
     "eslint": "^9.39.2",
-    "eslint-config-next": "^15.4.6",
+    "eslint-config-next": "^16.1.6",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-mdx": "^3.6.2",
     "eslint-plugin-react": "^7.37.5",
